requirements.txt
google-cloud-bigquery==3.25.0
google-cloud-storage==3.0.0
openpyxl==3.1.5
pandas==2.2.2
pyarrow==17.0.0
xlrd==2.0.1
gcsfs==2025.3.2